+++ CHANGES.rst
@@ -5,6 +5,10 @@ Changelog
 Unreleased
 ----------
 
+* Implemented rollback framework for CrateDB cluster operations. Introduced
+  ``RollbackHandler`` and specialized subhandler ``RollbackUpgradeSubHandler``
+  to support targeted rollback logic.
+
 2.49.0 (2025-07-14)
 -------------------